pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wallpaper-pipeline"
version = "0.2.0"
description = "Production-grade AI wallpaper compositor optimized for Windows with 32GB RAM - precise wall detection, object separation, and realistic lighting transfer"
authors = [{name = "ML Engineer", email = "engineer@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["wallpaper", "compositor", "ai", "computer-vision", "segmentation", "windows", "gpu"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    # Core ML Framework - Optimized for Windows with CUDA
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    
    # Computer Vision and Image Processing
    "opencv-python>=4.8.1",
    "opencv-contrib-python>=4.8.1",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "scikit-image>=0.21.0",
    
    # Segmentation Models
    "segmentation-models-pytorch>=0.3.3",
    "timm>=0.9.12",
    
    # Hugging Face Models
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "accelerate>=0.24.0",
    
    # Data Augmentation and Processing
    "albumentations>=1.3.1",
    
    # Visualization and Progress
    "matplotlib>=3.7.2",
    "tqdm>=4.66.0",
    
    # Windows-specific optimizations
    "psutil>=5.9.0",
    "memory-profiler>=0.61.0",
    
    # Performance and Caching
    "joblib>=1.3.0",
    "cachetools>=5.3.0",
]

[project.scripts]
wallpaper-pipeline = "src.scripts.run_batch:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
